Run shader model 4 too, and fix the six things that found
The counterpart machine for DXBC. A register here is four untyped words -
whether a word is a float, an int or a uint is the instruction's business,
not the register's - so the file is unsigned and every operation
reinterprets. Constants still come from the constant table by name, but
typed: a float bit pattern read as a uint is enormous, and a loop counting
up to one of those does not finish.

Six defects, all in the instruction writer, all of them compiling.

Saturate was dropped everywhere. It is a bit on the instruction rather than
anything in the operands, and nothing read it, so dp3_sat feeding a log
became log2 of a negative number. Every case assigns in the one shape, so
the clamp goes on around the expression.

An immediate took the first N components rather than the ones the
destination mask selects - the same defect as the D3D9 defined constant,
here in the other writer. `mov o0.zw, l(0, 0, 1, 2)` wrote 0 and 0, and
`mad o0.yz, v0.xy, l(0, 2, 2, 0), l(0, -1, -1, 0)` multiplied by 0.

The resource swizzle was dropped: `sample r0, v0.xyxx, t2.yzxw, s0` puts
green in x, and the output read the channels in order. The machine could
not see this at first because it ignored the swizzle as well - a harness
blind in the same place as the thing it checks will agree with it for the
wrong reason - so it applies it now.

ftoi, ftou, itof and utof were plain moves. Again the defect the AST writer
had, again surviving in the writer nothing compared: a temp is declared
with one integer type, so utof on a register holding 0xFFFFFFFF read it as
-1 where the original reads 4294967295.

ushr was written as >>, and HLSL reads >> as arithmetic or logical from the
type of what is shifted. The comment on ShiftRightOperation said so and the
code did it anyway, so a hash came out as its own complement. The unsigned
form now says uint at the value, in both writers.

One operand can read two packed scalars at once. `imax r0.yz, cb0[0].xy,
-cb0[0].xy` reads a and b, and naming the operand from its first component
read a twice.

Also decodes the test-boolean bit, so the disassembly can say if_z rather
than assuming if_nz. fxc always inverts the comparison instead, so no
fixture has one; the machine honours it rather than assuming.

172 of 173 compare, the one being the known cmp difference, and nothing is
skipped quietly.

+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/immediate_constant_buffer.fx
@@ -24,7 +24,7 @@ float4 main(float2 texcoord : TEXCOORD) : SV_Target
 	while (true) {
 		r2.x = (r0.w >= 4) ? -1 : 0;
 		if (r2.x != 0) break;
-		r0.z = r0.w;
+		r0.z = (float)(int)r0.w;
 		r2 = layers.Sample(samp, r0.xyz);
 		r0.z = dot(weights, icb[r0.w]);
 		r1 = r2 * r0.z + r1;
