+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/int_arithmetic.fx
@@ -6,7 +6,7 @@ float4 main(float4 texcoord : TEXCOORD) : SV_Target
 
 	int r0;
 	r0 = 3 * n + -7;
-	r0 = r0.x;
+	r0 = (float)(int)r0.x;
 	o = r0.x * texcoord;
 
 	return o;
