+++ HlslDecompiler.Tests/ShaderSources/vs_4_0_instruction/vertex_id.fx
@@ -12,9 +12,9 @@ float4 main(VS_IN i) : SV_Position
 
 	float4 r0;
 	float r1;
-	r0.x = i.sv_instanceid.x;
+	r0.x = (float)(uint)i.sv_instanceid.x;
 	r0 = r0.x * arr[1];
-	r1 = i.sv_vertexid.x;
+	r1 = (float)(uint)i.sv_vertexid.x;
 	o = arr[0] * r1.x + r0;
 
 	return o;
